Move logout button to UserProfile
